Added External and Local Fetch API's
It was added 2 Fetch API's in JSON format to push from external and local sources, making the resources in "numeros-grid" be fetched by a JSON archive in the main branch. And, in the footer a bitcoin value counter has added from a external Fetch API.

Features add with study purposes.